+++ yandex/resource_yandex_mdb_kafka_user_test.go
@@ -162,7 +162,6 @@ resource "yandex_mdb_kafka_cluster" "foo" {
 	  version          = "3.3"
 	  brokers_count    = 1
 	  zones            = ["ru-central1-a"]
-	  unmanaged_topics = true
 	  kafka {
 		resources {
 		  resource_preset_id = "s2.micro"